[-] Inline [Emit].

+++ Katrina/Executor/CodeExecutor.cs
@@ -21,7 +21,6 @@ namespace AV.Cyclone.Katrina.Executor
         private readonly Dictionary<CSharpCompilation, CSharpCompilation> compilations = 
             new Dictionary<CSharpCompilation, CSharpCompilation>();
         private List<CompilationEmitResult> compilationEmitResults;
-        private string tempDir;
 
         public void AddCompilation(CSharpCompilation oldCompilation, CSharpCompilation newCompilation)
         {
@@ -31,28 +30,6 @@ namespace AV.Cyclone.Katrina.Executor
             compilations[oldCompilation] = newCompilation;
         }
 
-        public void Emit()
-        {
-            // TODO: Delete emit
-            tempDir = Path.Combine(Path.GetTempPath(), "_Cyclon_" + Guid.NewGuid());
-            Directory.CreateDirectory(tempDir);
-            File.Copy(typeof(AssemblyLoader).Assembly.Location, Path.Combine(tempDir, typeof(AssemblyLoader).Assembly.GetName().Name + ".dll"));
-
-            compilationEmitResults = new List<CompilationEmitResult>(compilations.Count);
-
-            foreach (var compilation in compilations.Values)
-            {
-                var assemblyPath = Path.Combine(tempDir, compilation.AssemblyName + ".dll");
-                // TODO: Store emit results
-                var emitResult = compilation.Emit(assemblyPath);
-                compilationEmitResults.Add(new CompilationEmitResult
-                {
-                    AssemblyPath = assemblyPath,
-                    EmitResult = emitResult
-                });
-            }
-        }
-
         public void SetExecuteLogger(IExecuteLogger executeLogger)
         {
             Context.ExecuteLogger = executeLogger;
@@ -60,24 +37,61 @@ namespace AV.Cyclone.Katrina.Executor
 
         public void Execute(string compilationName, string className, string methodName)
         {
-            var executorDomain = AppDomain.CreateDomain("ExecutorDomain", AppDomain.CurrentDomain.Evidence, new AppDomainSetup {ApplicationBase = tempDir});
-            var loader = (AssemblyLoader)executorDomain.CreateInstanceAndUnwrap(typeof(AssemblyLoader).Assembly.FullName, typeof(AssemblyLoader).FullName);
-
-            int classAssemblyIndex = -1;
-            for (int i = 0; i < compilationEmitResults.Count; i++)
+            string tempDir = null;
+            AppDomain executorDomain = null;
+            try
             {
-                var compilationEmitResult = compilationEmitResults[i];
-                var assemblyName = AssemblyName.GetAssemblyName(compilationEmitResult.AssemblyPath);
-                loader.LoadAssembly(assemblyName);
-                if (assemblyName.Name == compilationName)
-                    classAssemblyIndex = i;
-            }
+                tempDir = Path.Combine(Path.GetTempPath(), "_Cyclon_" + Guid.NewGuid());
+                Directory.CreateDirectory(tempDir);
+                File.Copy(typeof (AssemblyLoader).Assembly.Location,
+                    Path.Combine(tempDir, typeof (AssemblyLoader).Assembly.GetName().Name + ".dll"));
 
-            loader.SetExecuteLogger(new DomainExecuteLogger(Context.ExecuteLogger));
-            loader.Execute(classAssemblyIndex, className, methodName);
-            AppDomain.Unload(executorDomain);
-            var tempDirInfo = new DirectoryInfo(tempDir);
-            tempDirInfo.Delete(true);
+                compilationEmitResults = new List<CompilationEmitResult>(compilations.Count);
+
+                foreach (var compilation in compilations.Values)
+                {
+                    var assemblyPath = Path.Combine(tempDir, compilation.AssemblyName + ".dll");
+                    // TODO: Store emit results
+                    var emitResult = compilation.Emit(assemblyPath);
+                    compilationEmitResults.Add(new CompilationEmitResult
+                    {
+                        AssemblyPath = assemblyPath,
+                        EmitResult = emitResult
+                    });
+                }
+
+                executorDomain = AppDomain.CreateDomain("ExecutorDomain", AppDomain.CurrentDomain.Evidence,
+                    new AppDomainSetup {ApplicationBase = tempDir});
+                var loader =
+                    (AssemblyLoader)
+                        executorDomain.CreateInstanceAndUnwrap(typeof (AssemblyLoader).Assembly.FullName,
+                            typeof (AssemblyLoader).FullName);
+
+                int classAssemblyIndex = -1;
+                for (int i = 0; i < compilationEmitResults.Count; i++)
+                {
+                    var compilationEmitResult = compilationEmitResults[i];
+                    var assemblyName = AssemblyName.GetAssemblyName(compilationEmitResult.AssemblyPath);
+                    loader.LoadAssembly(assemblyName);
+                    if (assemblyName.Name == compilationName)
+                        classAssemblyIndex = i;
+                }
+
+                loader.SetExecuteLogger(new DomainExecuteLogger(Context.ExecuteLogger));
+                loader.Execute(classAssemblyIndex, className, methodName);
+            }
+            finally
+            {
+                if (executorDomain != null)
+                {
+                    AppDomain.Unload(executorDomain);
+                }
+                if (!string.IsNullOrEmpty(tempDir))
+                {
+                    var tempDirInfo = new DirectoryInfo(tempDir);
+                    tempDirInfo.Delete(true);
+                }
+            }
         }
     }
 }
